Remove previous focus algorithm
