+++ src/view/components/profiler/flamegraph/ranked/ranked-utils.ts
@@ -1,4 +1,4 @@
-import { NodeTransform } from "../transform/focusNode";
+import { NodeTransform } from "../transform/shared";
 import { ProfilerNode, CommitData } from "../../data/commits";
 import { getGradient } from "../../data/gradient";
 import { ID } from "../../../../store/types";
